Authorize with GitHub.

This function is run automatically to allow gistr to access your GitHub account.

gist_auth(app = gistr_app, reauth = FALSE)

Arguments

app
An httr::oauth_app() for GitHub. The default uses an application gistr_oauth created by Scott Chamberlain.

reauth
(logical) Force re-authorization?

Value

a named list, with a single slot for Authorization, with a single element with the token - this is the expected format needed when passed down to the http request

Details

There are two ways to authorise gistr to work with your GitHub account:

Generate a personal access token with the gist scope selected, and set it as the GITHUB_PAT environment variable per session using Sys.setenv or across sessions by adding it to your .Renviron file or similar. See https://help.github.com/articles/creating-an-access-token-for-command-line-use for help

Interactively login into your GitHub account and authorise with OAuth.

Using GITHUB_PAT is recommended.

Examples

if (FALSE) {
gist_auth()
}